HSEARCH-4140 Move OutboxEvent-related code to a distinct package
