+++ integrationtest/mapper/orm-coordination-database-polling/src/test/java/org/hibernate/search/integrationtest/mapper/orm/automaticindexing/coordination/databasepolling/DatabasePollingAutomaticIndexingLifecycleIT.java
@@ -16,7 +16,7 @@ import javax.persistence.Id;
 
 import org.hibernate.SessionFactory;
 import org.hibernate.search.integrationtest.mapper.orm.coordination.databasepolling.FilteringOutboxEventFinder;
-import org.hibernate.search.mapper.orm.coordination.databasepolling.impl.OutboxEvent;
+import org.hibernate.search.mapper.orm.coordination.databasepolling.event.impl.OutboxEvent;
 import org.hibernate.search.mapper.pojo.mapping.definition.annotation.GenericField;
 import org.hibernate.search.mapper.pojo.mapping.definition.annotation.Indexed;
 import org.hibernate.search.util.impl.integrationtest.common.rule.BackendMock;
