+++ mapper/orm-coordination-database-polling/src/main/java/org/hibernate/search/mapper/orm/coordination/databasepolling/event/impl/DatabasePollingOutboxEventAdditionalJaxbMappingProducer.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.mapper.orm.coordination.databasepolling.impl;
+package org.hibernate.search.mapper.orm.coordination.databasepolling.event.impl;
 
 import java.io.BufferedInputStream;
 import java.io.ByteArrayInputStream;
@@ -27,7 +27,7 @@ import org.hibernate.search.util.common.logging.impl.LoggerFactory;
 import org.jboss.jandex.IndexView;
 
 @SuppressWarnings("deprecation")
-public class DatabasePollingOutboxEventAdditionalJaxbMappingProducer
+public final class DatabasePollingOutboxEventAdditionalJaxbMappingProducer
 		implements org.hibernate.boot.spi.AdditionalJaxbMappingProducer {
 
 	private static final Log log = LoggerFactory.make( Log.class, MethodHandles.lookup() );
