+++ mapper/orm-coordination-database-polling/src/main/java/org/hibernate/search/mapper/orm/coordination/databasepolling/event/impl/OutboxEventFinderProvider.java
@@ -4,12 +4,12 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.mapper.orm.coordination.databasepolling.impl;
+package org.hibernate.search.mapper.orm.coordination.databasepolling.event.impl;
 
-import java.util.List;
+import java.util.Optional;
 
-import org.hibernate.Session;
+public interface OutboxEventFinderProvider {
 
-public interface OutboxEventFinder {
-	List<OutboxEvent> findOutboxEvents(Session session, int maxResults);
+	OutboxEventFinder create(Optional<OutboxEventPredicate> predicate);
+
 }
